clean up internals
this patch optimizes the code for readability and clean interfaces.
mainly:
- separate functionality of `Board.get_area()` and
`Board.get_enclosed_area()` and clean up internally.
- remove pubsub dependency. now change of game state is returned to the
caller of `Game.command()` who can pass it on.
- reduce redundancy in computing player state. now everything happens in
`Game.update_player()` so it has not to be redone outside. `Player` thus
carries a bit more state.
- some more clarifying comments

hopefully the structure is now easier to understand and the program
operates a little more efficiently.